JAR hell solved
Runs, but takes forever to learn the 50 entries (reduced from 5000). Perhaps use
C or a different library entirely, perhaps one that uses the GPU? Maybe the
network is being too hard on itself and the error tolerance is too low. The library
might have an option to control the precision.